scripts/requirements.txt
pymysql
PyPDF2
eyed3
webdavclient3 
pillow
pymupdf